+++ testing/simulated/simulated_test.go
@@ -27,9 +27,11 @@ import (
 	"time"
 
 	"github.com/berachain/beacon-kit/beacon/blockchain"
+	"github.com/berachain/beacon-kit/consensus/cometbft/service/encoding"
 	"github.com/berachain/beacon-kit/log/phuslu"
 	"github.com/berachain/beacon-kit/primitives/common"
 	"github.com/berachain/beacon-kit/primitives/constants"
+	"github.com/berachain/beacon-kit/primitives/math"
 	"github.com/berachain/beacon-kit/testing/simulated"
 	"github.com/cometbft/cometbft/abci/types"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
@@ -42,9 +44,8 @@ const finalizeWaitDuration = 500 * time.Millisecond
 
 type Simulated struct {
 	suite.Suite
-	simulated.TestSuiteHandle
-	SimComet          *simulated.SimComet
-	BlockchainService *blockchain.Service
+	simulated.SharedAccessors
+	SimComet *simulated.SimComet
 	// LogBuffer gives us a mechanism to access the reason for a comet rejection
 	LogBuffer             *bytes.Buffer
 	GenesisValidatorsRoot common.Root
@@ -99,12 +100,6 @@ func (s *Simulated) SetupTest() {
 	s.NotNil(noopCometService)
 	s.SimComet = noopCometService
 
-	var blockchainService *blockchain.Service
-	err = testNode.FetchService(&blockchainService)
-	s.Require().NoError(err)
-	s.NotNil(blockchainService)
-	s.BlockchainService = blockchainService
-
 	go func() {
 		// Node blocks on Start and hence we have to run in separate routine
 		if err = s.TestNode.Start(s.Ctx); err != nil {
@@ -124,6 +119,7 @@ func (s *Simulated) TearDownTest() {
 }
 
 func (s *Simulated) TestFullLifecycle_ValidBlock_IsSuccessful() {
+	const height = 1
 	appGenesis, err := genutiltypes.AppGenesisFromFile(
 		s.HomeDir + "/config/genesis.json",
 	)
@@ -145,7 +141,7 @@ func (s *Simulated) TestFullLifecycle_ValidBlock_IsSuccessful() {
 	pubkey, err := blsSigner.GetPubKey()
 	s.Require().NoError(err)
 	proposal, err := s.SimComet.Comet.PrepareProposal(s.Ctx, &types.PrepareProposalRequest{
-		Height:          1,
+		Height:          height,
 		Time:            time.Now(),
 		ProposerAddress: pubkey.Address(),
 	})
@@ -154,7 +150,7 @@ func (s *Simulated) TestFullLifecycle_ValidBlock_IsSuccessful() {
 
 	processResponse, err := s.SimComet.Comet.ProcessProposal(s.Ctx, &types.ProcessProposalRequest{
 		Txs:    proposal.Txs,
-		Height: 1,
+		Height: height,
 		// If incorrect proposer address is used, we get a proposer mismatch error
 		ProposerAddress: pubkey.Address(),
 	})
@@ -165,7 +161,7 @@ func (s *Simulated) TestFullLifecycle_ValidBlock_IsSuccessful() {
 
 	finalizeResponse, err := s.SimComet.Comet.FinalizeBlock(s.Ctx, &types.FinalizeBlockRequest{
 		Txs:             proposal.Txs,
-		Height:          1,
+		Height:          height,
 		ProposerAddress: pubkey.Address(),
 	})
 	s.Require().NoError(err)
@@ -175,6 +171,23 @@ func (s *Simulated) TestFullLifecycle_ValidBlock_IsSuccessful() {
 	s.Require().NoError(err)
 
 	// Post state checks
+	queryContext, err := s.SimComet.CreateQueryContext(height, false)
+	s.Require().NoError(err)
+	stateDB := s.TestNode.StorageBackend.StateFromContext(queryContext)
+	slot, err := stateDB.GetSlot()
+	s.Require().NoError(err)
+	s.Require().Equal(math.U64(height), slot)
+
+	fetchedHeader, err := stateDB.GetLatestBlockHeader()
+	s.Require().NoError(err)
+
+	proposedBlock, err := encoding.UnmarshalBeaconBlockFromABCIRequest(
+		proposal.Txs,
+		blockchain.BeaconBlockTxIndex,
+		s.TestNode.ChainSpec.ActiveForkVersionForSlot(height),
+	)
+	s.Require().NoError(err)
+	s.Require().Equal(proposedBlock.Message.GetHeader().GetBodyRoot(), fetchedHeader.GetBodyRoot())
 }
 
 func (s *Simulated) TestInitChain_InvalidChainID_MustError() {
