+++ testing/simulated/utils.go
@@ -1,0 +1,46 @@
+// SPDX-License-Identifier: BUSL-1.1
+//
+// Copyright (C) 2025, Berachain Foundation. All rights reserved.
+// Use of this software is governed by the Business Source License included
+// in the LICENSE file of this repository and at www.mariadb.com/bsl11.
+//
+// ANY USE OF THE LICENSED WORK IN VIOLATION OF THIS LICENSE WILL AUTOMATICALLY
+// TERMINATE YOUR RIGHTS UNDER THIS LICENSE FOR THE CURRENT AND ALL OTHER
+// VERSIONS OF THE LICENSED WORK.
+//
+// THIS LICENSE DOES NOT GRANT YOU ANY RIGHT IN ANY TRADEMARK OR LOGO OF
+// LICENSOR OR ITS AFFILIATES (PROVIDED THAT YOU MAY USE A TRADEMARK OR LOGO OF
+// LICENSOR AS EXPRESSLY REQUIRED BY THIS LICENSE).
+//
+// TO THE EXTENT PERMITTED BY APPLICABLE LAW, THE LICENSED WORK IS PROVIDED ON
+// AN “AS IS” BASIS. LICENSOR HEREBY DISCLAIMS ALL WARRANTIES AND CONDITIONS,
+// EXPRESS OR IMPLIED, INCLUDING (WITHOUT LIMITATION) WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, NON-INFRINGEMENT, AND
+// TITLE.
+
+package simulated
+
+import (
+	"context"
+	"path/filepath"
+
+	"github.com/berachain/beacon-kit/node-core/components/signer"
+	"github.com/ory/dockertest"
+)
+
+// SharedAccessors holds references to common utilities required in tests
+type SharedAccessors struct {
+	Ctx        context.Context
+	CancelFunc context.CancelFunc
+	HomeDir    string
+	TestNode   TestNode
+
+	// EL dockertest handles for closing
+	ElHandle *dockertest.Resource
+}
+
+func GetBlsSigner(tempHomeDir string) *signer.BLSSigner {
+	privValKeyFile := filepath.Join(tempHomeDir, "config/priv_validator_key.json")
+	privValStateFile := filepath.Join(tempHomeDir, "data/priv_validator_state.json")
+	return signer.NewBLSSigner(privValKeyFile, privValStateFile)
+}
